Add failing spec to demonstrate issue with attribs
When a record is fetched from the server, any attributes the server
returns that are not declared on the Spine model are preserved and
accessible on the client side. Examples might include Rails' :created_at
and :updated_at fields.

However, if a record is created client-side and saved to the server,
Spine does not allow those additional undeclared attributes to get set
on the record.